# Bug fixes and PartitionedTable constituent optimization work