--- a/Integrations/python/deephaven_legacy/java_to_python.py
+++ b/Integrations/python/deephaven_legacy/java_to_python.py
@@ -545,7 +545,7 @@ def tableToDataFrame(table, convertNulls=NULL_CONVERSION.ERROR, categoricals=Non
     convertNulls = NULL_CONVERSION.validateValue(convertNulls)
 
     # get the column objects
-    columnDefs = table.getDefinition().getColumns()
+    columnDefs = table.getDefinition().getColumnsArray()
     # extract the names
     columnNames = [col.getName() for col in columnDefs]
     # NB: map will provide a generator in Python 3, which will fail here...