--- a/Util/src/main/java/io/deephaven/util/datastructures/SimpleReferenceManager.java
+++ b/Util/src/main/java/io/deephaven/util/datastructures/SimpleReferenceManager.java
@@ -40,8 +40,20 @@ public final class SimpleReferenceManager<T, R extends SimpleReference<T>> {
      * @param concurrent Use CopyOnWriteArrayList for internal storage if true, else ArrayList
      */
     public SimpleReferenceManager(@NotNull final Function<T, R> referenceFactory, final boolean concurrent) {
+        this (referenceFactory, concurrent ? new CopyOnWriteArrayList<>() : new ArrayList<>());
+    }
+
+    /**
+     * Create a SimpleReferenceManager with the specified {@code referenceCollection} as the backing data structure..
+     *
+     * @param referenceFactory Factory to create references for added referents; should always make a unique reference
+     * @param referenceCollection The collection to use for holding references
+     */
+    public SimpleReferenceManager(
+            @NotNull final Function<T, R> referenceFactory,
+            @NotNull final Collection<R> referenceCollection) {
         this.referenceFactory = referenceFactory;
-        references = concurrent ? new CopyOnWriteArrayList<>() : new ArrayList<>();
+        this.references = referenceCollection;
     }
 
     /**
@@ -224,7 +236,7 @@ public final class SimpleReferenceManager<T, R extends SimpleReference<T>> {
         // This is particular to the use case in question, because the deque is used as an ordered subset of an ordered
         // data structure for removing items by identity match. Don't try to use this deque for anything else and expect
         // you'll like the results.
-        return new ArrayDeque<R>() {
+        return new ArrayDeque<>() {
             @Override
             public boolean contains(final Object object) {
                 if (peekFirst() == object) {