--- a/engine/api/src/main/java/io/deephaven/engine/table/Table.java
+++ b/engine/api/src/main/java/io/deephaven/engine/table/Table.java
@@ -1827,9 +1827,13 @@ public interface Table extends
     Table coalesce();
 
     /**
-     * Get a {@link Table} that contains a sub-set of the rows from {@code this}.
+     * Get a {@link Table} that contains a sub-set of the rows from {@code this}. The result will share the same
+     * {@link #getColumnSources() column sources} and {@link #getDefinition() definition} as this table.
      *
-     * @param rowSet The {@link TrackingRowSet row set} for the result.
+     * The result will not update on its own. The caller must also establish an appropriate listener to update
+     * {@code rowSet} and propagate {@link TableUpdate updates}.
+     *
+     * @param rowSet The {@link TrackingRowSet row set} for the result
      * @return A new sub-table
      */
     Table getSubTable(TrackingRowSet rowSet);