1. Adds PartitionedTable and friends to Groovy session imports
2. Fixes dependency check bug found in enterprise
3. Port query performance nugget elimination for `QueryTable.getSubTable` from enterprise
4. Retire direct listeners
5. Adapt shift-oblivious listeners to be shift-aware listeners
6. Fix bug that skipped clearing shift-aware listeners on `BaseTable.destroy`
7. Get rid of `Serializable` or `Externalizable` in several undesirable places
8. Eliminate many unnecessary uses of `DataColumn`
9. Switch most/all data structures used by `BaseTable` and `QueryTable` to be lazily initialized
10. Allow attribute sharing for `Table.partitionBy` sub-tables
11. Make `ColumnDefinition` and `TableDefinition` immutable
12. Allow interning for `TableDefinition` and make it the default for `QueryTable`

Fixes #2420