@rcaudy 2022-05-31
I ported a small unit test from @cpwright 's un-merged enterprise branch:

Current form:
```
    public void testPerformance() {
        final RuntimeMemory runtimeMemory = RuntimeMemory.getInstance();
        final RuntimeMemory.Sample startMemoryStats = new RuntimeMemory.Sample();
        final RuntimeMemory.Sample endMemoryStats = new RuntimeMemory.Sample();
        final Table table = TableTools.emptyTable(10_000_000).update("X=ii % 2_500_000").updateView("X2=ii");
        runtimeMemory.read(startMemoryStats);
        final long start = System.nanoTime();
        final PartitionedTable pt1 = table.partitionBy("X");
        final PartitionedTable pt2 = pt1.transform(t -> t.view("Y=X", "Y2=X2"));
        System.gc();
        System.gc();
        final long end = System.nanoTime();
        runtimeMemory.read(endMemoryStats);
        final long changeBytes = (endMemoryStats.totalMemory - endMemoryStats.freeMemory)
                - (startMemoryStats.totalMemory - startMemoryStats.freeMemory);
        System.out.println("Elapsed: " + ((end - start) / 1_000_000) + " ms, " + changeBytes + " bytes used");
    }
```

Performance numbers:
```
Just partitionBy:
Mine (no intern): 1834 ms, 1158507520 bytes used
                  1513 ms, 1130887168 bytes used
            Mine: 1504 ms, 1179438080 bytes used
                  1280 ms, 1178373144 bytes used
            Main: 7096 ms, 3047465472 bytes used
                  6738 ms, 2980094464 bytes used
                  6623 ms, 2974899200 bytes used
       No-nugget: 4419 ms, 3261266432 bytes used
                  3607 ms, 3202976768 bytes used

With view:
Mine (no-intern): 15941 ms, 3080512000 bytes used
                  19761 ms, 3162089472 bytes used
                  16528 ms, 2698629312 bytes used (with GC)
                  16401 ms, 2701340688 bytes used (with GC)
   Mine (intern): 16348 ms, 2627812864 bytes used
                  15549 ms, 2967764992 bytes used
                  15896 ms, 2304043736 bytes used (with GC)
                  15997 ms, 2291313280 bytes used (with GC)
                  16479 ms, 2293743160 bytes used (with GC)
            Main: OOM
```